+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Make `StoreKeyRange` and `StoreKeyStartValue` clonable
  - **Breaking**: Remove `ReadableWritableStorageTraits`, `ReadableWritableStorage`, `ReadableWritableStore`, and `StorageTransformerExtension::create_readable_writable_transformer`
    - These were redundant because `WritableStorageTraits` requires `ReadableStorageTraits` since `6e69a4d`
+ - Move sync stores to `storage::store::sync`
 
 ## [0.6.0] - 2023-11-16